Fix lint usage w/ prettier.
We were missing the `"prettier/prettier": "error"` rule which means that we disabled all style rules and _didn't_ have anything re-added for prettier to fix specifically. :P  This corrects the config and when run we get this (expected) error:

```
$ eslint .

/Users/rye/scm/fmd/nuka-carousel/src/index.js
  1028:46  error  Insert `;`  prettier/prettier
  1029:46  error  Insert `;`  prettier/prettier
  1030:46  error  Insert `;`  prettier/prettier
```

And `eslint . --fix` or `yarn lint --fix` now automagically fixes prettier-based errors!