Prevent disabling click during neglible mouse movement
When dragging is true, it has been noted that users may find themselves trying to click on an image and nothing happens. This is because without them knowing, `onMouseMove` has occured with negligible movement to the user, but noticeable by the code which therefore disables clicking for the user. This is not optimal UX.

"10px" was chosen because this gives a user some leeway when they are trying to click something while accidentally barely moving the slide.